Added utility for moving user profiles into private image.
move-profiles needs to run before any user is logged on.
It takes xen/vbd ID of the private disk image device as parameter.
It initializes/partitions/formats the disk if needed, copies the
profiles directory there, deletes the original and replaces it
with a directory junction point.
It's designed to run as a local group policy startup "script".
That makes it run before services and blocks logon until completion.

bootstrap is the actual process that runs as a startup script.
It in turn spawns move-profiles. That's because for some reason
startup scripts seem to lack many privileges when running on
Qubes Windows VMs, even if they do run as LocalSystem.
bootstrap ensures that the spawned process has the correct privileges,
especially SeCreateSymbolicLinkPrivilege.