Add qubes.StartApp service
This service is used in Qubes 4.0 to start application from the menu.
Service is given application name as an argument - the name is the same
as a base name sent by qubes.GetAppmenus service.
Since the encoding there is non-reversible (spaces and backslashes are
replaced by _ and -, but original _ and - are not escaped), keep a
mapping in registry - similar to the one for retrieving icons.

QubesOS/qubes-issues#3585